First part of changing the domain.
