+++ scattersphere-core/pom.xml
@@ -3,7 +3,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>com.suenobu</groupId>
+        <groupId>io.buildfactory</groupId>
         <artifactId>scattersphere</artifactId>
         <version>0.2.1</version>
         <relativePath>..</relativePath>
