--- scattersphere-core/src/main/scala/com/scattersphere/core/util/execution/package.scala
@@ -1,9 +1,0 @@
-package com.scattersphere.core.util
-
-/** Provides classes to execute [[Job]]s.
-  *
-  * For an overview of [[Job]]s and [[Task]]s, please review [[com.scattersphere.core.util]] classes.
-  *
-  * @since 0.0.1
-  */
-package object execution { }
