+++ scattersphere-core/src/main/scala/io/buildfactory/scattersphere/core/util/execution/package.scala
@@ -1,0 +1,9 @@
+package io.buildfactory.scattersphere.core.util
+
+/** Provides classes to execute [[Job]]s.
+  *
+  * For an overview of [[Job]]s and [[Task]]s, please review [[io.buildfactory.scattersphere.core.util]] classes.
+  *
+  * @since 0.0.1
+  */
+package object execution { }
